// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const FileType = {
  FILE: 'FILE',
  FOLDER: 'FOLDER'
} as const

export type FileType = (typeof FileType)[keyof typeof FileType]